+++ SSR_Collaboration_BU/SSR_Collaboration_BU/MyCircle.cpp
@@ -1,11 +1,11 @@
 #include"DxLib.h"
 #include"MyCircle.h"
 
-MyCircle::MyCircle(double i_r):r(i_r) {}
+MyCircle::MyCircle(float i_r):r(i_r) {}
 
 MyCircle::~MyCircle() {}
 
-void MyCircle::VDraw(double x, double y, unsigned int color, int fillFlag, int textureHandle) {
+void MyCircle::VDraw(float x, float y, unsigned int color, int fillFlag, int textureHandle) {
 	//ひとまずテクスチャーを使わずに実装
 	DrawCircle((int)x, (int)y ,(int)r, color, fillFlag);
 }
